refactor: move helper functions to helper module instead of conftest.py
